+++ Realm/RLMRealmConfiguration.mm
@@ -66,7 +66,7 @@ NSString *RLMRealmPathForFile(NSString *fileName) {
 
 @implementation RLMRealmConfiguration {
     realm::Realm::Config _config;
-    RLMSyncConfiguration *_syncConfiguration;
+    RLMSyncErrorReportingBlock _manualClientResetHandler;
 }
 
 - (realm::Realm::Config&)configRef {
@@ -362,8 +362,10 @@ static bool isSync(realm::Realm::Config const& config) {
     _config.in_memory = false;
     _config.sync_config = std::make_shared<realm::SyncConfig>(syncConfiguration.rawConfiguration);
     _config.path = syncConfiguration.path;
-    // Assign _syncConfiguration for RLMSyncConfiguration properties that do not exist on the raw sync_config.
-    _syncConfiguration = syncConfiguration;
+
+    // The manual client reset handler doesn't exist on the raw config,
+    // so assign it here.
+    _manualClientResetHandler = syncConfiguration.manualClientResetHandler;
 
     [self updateSchemaMode];
 }
@@ -373,7 +375,7 @@ static bool isSync(realm::Realm::Config const& config) {
         return nil;
     }
     RLMSyncConfiguration* syncConfig = [[RLMSyncConfiguration alloc] initWithRawConfig:*_config.sync_config path:_config.path];
-    syncConfig.manualClientResetHandler = _syncConfiguration.manualClientResetHandler;
+    syncConfig.manualClientResetHandler = _manualClientResetHandler;
     return syncConfig;
 }